[refactor] Contribute => many split, cleaning, copy/paste.
The page "Contribute" on the website was modified into a "Community" page ; with subcategories: the communities, community project ( derivations ) , fan-arts, and then under the "how to contribute". It was also a perfect reason to do cleaning and cut all paragraphs into their independant block ; it will be useful later if the layout become a bit more dynamic.