Remove URL in comments for author names
This commit removes the URL because a lot of spammer use it to let totally normal comments and hide behind the author name clickable an awfull URL that can put pepper&carrot in a bad network. My interface to review comments doesn't display this information in backend ; so someone can always send a message with hidding a very bad URL. I also removed the field optional to enter a website URL adress. It should help.