Fix #8: Add Terms of service and Privacy for GDPR (General Data Protection Regulation) compliance.
This add a new file at root TERMS-OF-SERVICE-AND-PRIVACY.md.
Proofreading might be necessary from a skilled English writer.

The fix also add a new static page to display static markdown file
and a CSS class to style this document the official way and ease
reading.

A bit of cleanup was done on the footer.php template, to display
social network icons, and also the cleanup was reflected into the
header while I was doing that.