[refactor] Fix #11 Language pills display on homepage and webcomics
This commit mix a bit the translation system of the website and of the
webcomic. So far, I kept them a bit not mixed. I'm sure this change is
creating odd UX (eg. clicking an uncomplete translated lang redirect to
the only page 100% sure translated: the episode 01) but no one on #11
had a better idea about what happens when someone click a pills with
the active page not translated. For sure the previous situation was
certainly more logic.

So, what is the benefit of this commit?
- Discoverability of language (eg. new visitor discover we also provide
Latin, Sinhalese, etc; direcly from the Homepage or the webcomic page)

Note: This change affect only Homepage and Webcomic page. For all other
content, it wasn't making a lot of sens.