[big-merge][post-fix] Don't try to guess lang from URL, it doesn't work in frontpage...
The frontpage in PluXML is really particular; and it makes the code
of the webcomic menu confused. The URL now is www.peppercarrot.com;
that's the convention for every 'homepage' defined by PluXML, the
URL of root, and nothing more.

I wish I had peppercarrot.com/fr/ or something similar. Easier to
share the homepage with a lang associated to it.

Anyway, without this commit; the page was tempted to load no comics.
(no detection in URL == lang is not existing ). Now the script tries
to look in other places.