Refactor "community" entry in the top menu
Ref #29

Now the community page displays 'cards'; this one are generated
from a base of *.json files stored (temporarly) under community
folder at root of the folder.

This commit remove the last dependencies to blog post thumbnails
(before this commit, the community page still listed all blog
post of derivations).

The wiki moved from the top menu into a card. This will provide
room to add a 'eshop' soon™.

I hope this change will bring more visibility to the E-Reader,
P&Cmini and the link to the source of various projects.