Refactor the cards: theme using gfx sprite + new way to get called
The new way to get called works around a bug where the translation
system couldn't access the content of the previous function. Making
it this way works perfect and I even felt it more logical and easier.

The new theme works now with image-borders; inspired by my work on the
wiki.