Refactor Sources to adopt new theme
The code base of 'Sources' is complex and this is a lazy refactor;
I only scratched the surface and decorated the static elements.

But other parts are working correctly; so why fix it? This is still
a very optional 'backend' of the webcomic experience. It is still
better than just giving the 0_sources folder. (as many FLOSS project
does).