Remove a sentence for the eShop
After a night on it this one felt wrong: I can't say 'worldwide
shipping': it depends the platform. I can't say also benefits
goes to the author: as a part goes to the platform too... And for
"nice price" it might be offensive for some countries where the
conversion to their device makes the price pretty high and not nice.

Well, better to remove this sentence. It is one string less to
translate.