[Sources/Downloader] Refactor the user interface (with seasons)
The sources/downloader now use the JSON to validate a lang and
display a full name of the lang in the list.

A new option is added: a season selector. This send a $s variable
to the downloader.php script; allowing later to produce smaller zip
files (actual big zip files starts to kill the server at each
generation...)

Also, wording, better comments for the code, etc...