Add better UX, anti flood system and CPU/friendly rules.
The previous download.php script was really intensive for the server,
this new version adds a system to limit to 8 zip generated per days,
ensure 3min is spent in-between the generations, and adds more checks
to not generate random files from malformed queries.

It also now log on success and have a better user interface.