Allows the sources to read "new-episodes" and propose a XYZ preview
Updating XYZ is really long and a manual process. By doing this
change, I'm allowing the Sources page of the website to display
the future episode (it has the new- prefix). So the new episode
benefits now from all automation already done for regular
translation. It unify the workflow on my side greatly.