Add link on XYZ to https://peppercarrot.frama.io/gitlab-renderer/
This is the renderer of Merge Request made by @valvin, a very
cool and useful feature.

In the future, I want XYZ to be plugged into "Sources" more
easily, with a simplier name (Beta-test area) and with all of
that easy to find for the contributors. Right now, things are
still a bit in the proof-of-concept html trown together on a
page. Ergonomy will come.