[credit generator feature, v1a] Generate a long credit list
This list was made to help at generating ever-green/always fresh
credits. In fact, I'm working on the book project now and I realised
that all my work on the credits in November (that took me days) were
obsolete and I had to start over all again.

Fortunately, the work of @gunchleoc on the credit/data extraction
repository wide offered an opportunity to automatize and generate that.

So this script is the first version.

TODO I might do later:
- Removing too many "editing" and replace it with "contribution" in
epXX-XXX/lang/xx/info.json for the one who contributed to a lang without
necessarly "proofreading" (title art/speechbubble layout improvement).
General touch of SVG or maintainance is now at the root of the project.
- Find a system to detect the original language and remove "translation"
in this case and show "-original language-" label. Maybe use this data
from the info.json at the root of episode or detect the one in the language
if the translation tag has the "original" or "" keywords (it sounds less
reliable on long term). But because many translator duplicate English folder
and this one being the original language 90% of time now, it requires a
field ready to fill. So the "translation: none", "translation: original" is
a good candidate here.

Ref: https://framagit.org/peppercarrot/webcomics/-/issues/160

Demo: https://www.peppercarrot.com/en/static6/sources&page=credits