Better theme the a hyperlinks.
For long on this theme I used a green color for the link without
underline. I also many time removed all decoration for links because
it was looking a bit too buzy visualy.

I'm trying here to go back to something a big more classic: a teal/blue
desaturated with a brighter underline.I think it is still pleasant while
the color+underline makes it obvious this is an actionable hyperlink.

Ref: https://framagit.org/peppercarrot/webcomics/-/issues/160