[chatroom/contribute] Attempt at simplifying invitation to connect with us to the chat room
Before, the page had just three buttons without context. Here I tried to put focus on a
single button to our Element room (user click; a page appear with the log of the room and
propose to create an account to interact; this is super simple).
I also added the IRC information for advanced users and adress to the telegram bridge.
About Matermost, I think the bridge never really was active; so I abandoned the button.